requirements.txt
Django==1.6.1
dj-database-url==0.2.2
dj-static==0.0.5
django-nose==1.2
django-picklefield==0.3.1
django-social-auth==0.7.28
django-tinymce==1.5.2
django-toolbelt==0.0.1
gunicorn==18.0
httplib2==0.8
nose==1.3.0
oauth2==1.5.211
pil
psycopg2==2.5.2
pystache==0.5.3
python-openid==2.2.5
static==1.0.2
wsgiref==0.1.2

-e git+git://github.com/comoga/django-constance@dc89ec294f#egg=django-constance
-e git+git://github.com/kamni/django-null-reality#egg=django-null-reality
-e git+git://github.com/kamni/django-test-utils#egg=django-test-utils